requirements.txt
asgiref==3.2.7
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
Deprecated==1.2.9
Django==2.1
django-basicauth==0.5.2
djongo==1.3.2
dnspython==1.16.0
idna==2.9
PyGithub==1.47
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.9
wrapt==1.12.1
